Have "make bootstrap-local" do "make pull" first
In case the named (tagged) image is not available, the behavior of
docker-compose is to use "build" rather than pulling. Force the
opposite to happen through the Makefile, by ensuring that the tagged
image is indeed available before "make up".

N.B.: this problem only presents itself for new users that *do not*
yet have the images downloaded.