Hoist ExternalMenuItem::all() into model.php
This requires dodging a bullet: there are no true class variables in
PHP, or more precisely, these are inherited like methods. In order for
subclasses of TypedPost not to clobber each other's ::all() caches, we
could 1) demand that each subclass define its own variable (resulting
in a silent catastrophe in case that's forgotten) or 2) do as in this
commit i.e. key the cache by class name.