Correct homepage URLs (fixes wwp-1936)
Any place where the homepage of a site is included, there is some
weird magic going down to sort of canonicalize the URL of the home
page with respect to the language. Unfortunately, the homepage tries
hard to second-guess the language the visitor wants to be in, which
makes language-specific URLs kind of hard to make work.

* Do our part on the epfl-menus side: in order for stitching to work
  as designed, homepages must be language-specific links just like all
  the others